Refactor step3
