+++ src/main/java/nextstep/subway/line/domain/Line.java
@@ -21,8 +21,8 @@ public class Line extends BaseEntity {
     private String name;
     private String color;
 
-    @OneToMany(mappedBy = "line", cascade ={CascadeType.PERSIST, CascadeType.MERGE}, orphanRemoval = true)
-    private List<Section> sections = new ArrayList<>();
+    @Embedded
+    private Sections sections = new Sections();
 
     public Line() {
     }
@@ -49,95 +49,31 @@ public class Line extends BaseEntity {
         return color;
     }
 
-    public void addSection(Section section) {
-        if (isAlreadyRegistered(section) || !isAppendableSection(section) ) {
-            throw new InvalidSectionException("Input section is invalid");
-        }
-        this.sections.add(section);
-        section.setLine(this);
-    }
-
-    private Boolean isLeftOneSection(){
-        return (this.sections.size() == 1);
-    }
-
-    private Boolean isEmptySections() {
-        return (this.sections.size() == 0);
-    }
-
-    private Boolean isAppendableSection(final Section section) {
-        if (isEmptySections()) {
-            return true;
-        }
-        return (getLastDownStationId() == section.getUpStation().getId());
-    }
-
-    private Boolean isAlreadyRegistered(final Section section) {
-        return sections.stream()
-                .anyMatch(s -> s.getUpStation().equals(section.getDownStation())
-                        && s.getDownStation().equals(section.getDownStation()));
+    public void addSection(Station upStation, Station downStation, int distance) {
+        sections.add(new Section(this, upStation, downStation, distance));
     }
 
     public void deleteStation(final Station station){
-        final Section section = getLastSection();
-        if(!section.getDownStation().equals(station) || isLeftOneSection()){
-            throw new InvalidSectionException("Invalid Station Id" + station.getId());
-        }
-        sections.remove(section);
+        sections.delete(station);
     }
 
-    public List<StationResponse> getStations(){
-        List<Station> sequentialStations = new ArrayList<>();
-        for (Station station = firstStation(); station != null; station = nextStation(station)){
-            sequentialStations.add(station);
-        }
-        return sequentialStations.stream()
-                .map(StationResponse::of).collect(Collectors.toList());
-    }
-
-    private Station firstStation(){
-        return sections.stream()
-                .filter(section -> !matchAnyDownStation(section.getUpStation()))
-                .map(section -> section.getUpStation())
-                .findFirst()
-                .orElse(null);
-    }
-
-    private Station nextStation(final Station station) {
-        return sections.stream()
-                .filter(section -> section.getUpStation().equals(station))
-                .map(section -> section.getDownStation())
-                .findFirst()
-                .orElse(null);
-    }
-
-    public Section getLastSection(){
-        return sections.stream()
-                .filter(section -> !matchAnyUpStation(section.getDownStation()))
-                .findFirst()
-                .orElse(null);
-    }
-
-    private Boolean matchAnyUpStation(final Station station) {
-        return sections.stream()
-                .anyMatch(section -> section.getUpStation().equals(station));
-    }
-
-    private Boolean matchAnyDownStation(final Station station) {
-        return sections.stream()
-                .anyMatch(section -> section.getDownStation().equals(station));
+    public List<Station> listStations(){
+        return sections.getStations();
     }
 
     public Long getLastUpStationId(){
-        return getLastSection().getUpStation().getId();
+        return sections.getLastSection().getUpStation().getId();
     }
 
     public Long getLastDownStationId(){
-        return getLastSection().getDownStation().getId();
+        return sections.getLastSection().getDownStation().getId();
     }
 
-    public Integer getLineDistance() {
-        return sections.stream().
-                reduce(0, (TotalDistance, section) -> TotalDistance + section.getDistance(), Integer::sum);
+    public Integer getLineDistance(){
+        return sections.getLineDistance();
+    }
+
+    public Section getLastSection(){
+        return sections.getLastSection();
     }
 }
