+++ src/test/java/nextstep/subway/line/LineAcceptanceTest.java
@@ -1,19 +1,17 @@
 package nextstep.subway.line;
 
-import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import nextstep.subway.AcceptanceTest;
 import nextstep.subway.line.dto.CreatedLineResponse;
 import nextstep.subway.line.dto.LineResponse;
 import nextstep.subway.line.dto.SectionResponse;
-import nextstep.subway.station.StationAcceptanceTest;
+import nextstep.subway.station.StationSteps;
 import nextstep.subway.station.dto.StationResponse;
 import org.assertj.core.groups.Tuple;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
 
 import java.util.*;
 import java.util.stream.Collectors;
@@ -42,117 +40,21 @@ public class LineAcceptanceTest extends AcceptanceTest {
     void createLine() {
         // given
         // 지하철 노선
-        final Map newLine = createLineMapHelper("KangNam Line","green");
-        final Map upStation = StationAcceptanceTest.createStationInputHelper("강남역");
-        final Map downStation = StationAcceptanceTest.createStationInputHelper("서초역");
+        final Map newLine = LineSteps.createLineMapHelper("KangNam Line","green");
+        final Map upStation = StationSteps.createStationInputHelper("강남역");
+        final Map downStation = StationSteps.createStationInputHelper("서초역");
         final int distance = 10;
 
         // when
         // 지하철_노선_생성_요청
-        ExtractableResponse<Response> response = registerLineWithStationsHelper(newLine, upStation, downStation, distance);
+        ExtractableResponse<Response> response = LineSteps.registerLineWithStationsHelper(newLine, upStation, downStation, distance);
 
         // then
         // 지하철_노선_생성됨
         assertCreateNewLineSuccess(response);
     }
 
-    @DisplayName("구간을 추가한다.")
-    @Test
-    void createSection() {
-        // given
-        // 지하철 노선
-        CreatedLineResponse createdLineResponse  =
-                registerLineWithStationsHelper("KangName Line", "green", "강남역","서초역", 10)
-                .as(CreatedLineResponse.class);
 
-        final Map newDownStation = StationAcceptanceTest.createStationInputHelper("방배역");
-        StationResponse newDownStationResponse = StationAcceptanceTest.createStationHelper(newDownStation).as(StationResponse.class);
-
-        // when
-        Map<String, String> newSection = createSectionMapHelper(
-                createdLineResponse.getDownStationId(),
-                newDownStationResponse.getId(),
-                5);
-        ExtractableResponse<Response> newSectionResponse = registerSectionHelper(createdLineResponse.getId(), newSection);
-
-        // then
-        assertAppendNewSectionSuccess(newSectionResponse);
-    }
-
-    @DisplayName("구간을 제거한다.")
-    @Test
-    void deleteSection() {
-        // given
-        // 지하철 노선
-        // 구간 추가
-        CreatedLineResponse createdLineResponse  =
-                registerLineWithStationsHelper("KangName Line", "green", "강남역","서초역", 10)
-                        .as(CreatedLineResponse.class);
-
-        final Map newDownStation = StationAcceptanceTest.createStationInputHelper("방배역");
-        StationResponse newDownStationResponse = StationAcceptanceTest.createStationHelper(newDownStation).as(StationResponse.class);
-        Map<String, String> newSection = createSectionMapHelper(
-                createdLineResponse.getDownStationId(),
-                newDownStationResponse.getId(),
-                5);
-        SectionResponse newSectionResponse = registerSectionHelper(createdLineResponse.getId(), newSection).as(SectionResponse.class);
-
-        // when
-        // 구간 삭제
-        ExtractableResponse<Response> deleteSectionResponse = deleteStationHelper(createdLineResponse.getId(), newDownStationResponse.getId());
-
-        // then
-        assertDeleteSuccess(deleteSectionResponse);
-    }
-
-    @DisplayName("하나가 남은 구간은 제거가 안된다.")
-    @Test
-    void deleteLastOneSection() {
-        // given
-        // 지하철 노선
-        // 구간 추가
-        CreatedLineResponse createdLineResponse  =
-                registerLineWithStationsHelper("KangName Line", "green", "강남역","서초역", 10)
-                        .as(CreatedLineResponse.class);
-
-        final Map newDownStation = StationAcceptanceTest.createStationInputHelper("방배역");
-        StationResponse newDownStationResponse = StationAcceptanceTest.createStationHelper(newDownStation).as(StationResponse.class);
-        Map<String, String> newSection = createSectionMapHelper(
-                createdLineResponse.getDownStationId(),
-                newDownStationResponse.getId(),
-                5);
-        SectionResponse newSectionResponse = registerSectionHelper(createdLineResponse.getId(), newSection).as(SectionResponse.class);
-
-        // when
-        // 구간 삭제
-        ExtractableResponse<Response> deleteSectionResponse = deleteStationHelper(createdLineResponse.getId(), newSectionResponse.getId());
-
-        // then
-        assertBadRequestFail(deleteSectionResponse);
-    }
-
-    @DisplayName("등록하려는 구간의 상행역이 종점역이 아닌경우 등록되지 않는다.")
-    @Test
-    void createSectionWithNotMatchedDownStation() {
-        // given
-        // 지하철 노선
-        CreatedLineResponse createdLineResponse  =
-                registerLineWithStationsHelper("KangName Line", "green", "강남역","서초역", 10)
-                        .as(CreatedLineResponse.class);
-
-        final Map newDownStation = StationAcceptanceTest.createStationInputHelper("방배역");
-        StationResponse newDownStationResponse = StationAcceptanceTest.createStationHelper(newDownStation).as(StationResponse.class);
-
-        // when
-        Map<String, String> newSection = createSectionMapHelper(
-                createdLineResponse.getUpStationId(),
-                newDownStationResponse.getId(),
-                5);
-        ExtractableResponse<Response> newSectionResponse = registerSectionHelper(createdLineResponse.getId(), newSection);
-
-        // then
-        assertBadRequestFail(newSectionResponse);
-    }
 
 
     @DisplayName("지하철 노선 목록을 조회한다.")
@@ -162,15 +64,13 @@ public class LineAcceptanceTest extends AcceptanceTest {
         // 지하철_노선_등록되어_있음
         // 지하철_노선_등록되어_있음
         CreatedLineResponse kangNamLineResponse  =
-                registerLineWithStationsHelper("KangName Line", "green", "강남역","서초역", 10)
-                        .as(CreatedLineResponse.class);
+                LineSteps.registerLineWithStationsHelper("KangName Line", "green", "강남역","서초역", 10);
         CreatedLineResponse bunDangLineResponse  =
-                registerLineWithStationsHelper("Bundang Line", "yellow", "양재역","분당역", 20)
-                        .as(CreatedLineResponse.class);
+                LineSteps.registerLineWithStationsHelper("Bundang Line", "yellow", "양재역","분당역", 20);
 
         // when역
         // 지하철_노선_목록_조회_요청
-        ExtractableResponse<Response> response = getLinesHelper();
+        ExtractableResponse<Response> response = LineSteps.getLinesHelper();
 
         // then
         // 지하철_노선_목록_응답됨
@@ -185,12 +85,11 @@ public class LineAcceptanceTest extends AcceptanceTest {
         // given
         // 지하철_노선_등록되어_있음
         CreatedLineResponse registeredLine  =
-                registerLineWithStationsHelper("KangName Line", "green", "강남역","서초역", 10)
-                        .as(CreatedLineResponse.class);
+                LineSteps.registerLineWithStationsHelper("KangName Line", "green", "강남역","서초역", 10);
 
         // when
         // 지하철_노선_조회_요청
-        ExtractableResponse<Response> response = findLineHelper(registeredLine);
+        ExtractableResponse<Response> response = LineSteps.findLineHelper(registeredLine);
 
         // then
         // 지하철_노선_응답됨
@@ -203,19 +102,18 @@ public class LineAcceptanceTest extends AcceptanceTest {
         // given
         // 지하철_노선_등록되어_있음
         CreatedLineResponse registeredLine  =
-                registerLineWithStationsHelper("KangName Line", "green", "강남역","서초역", 10)
-                        .as(CreatedLineResponse.class);
+                LineSteps.registerLineWithStationsHelper("KangName Line", "green", "강남역","서초역", 10);
 
         // when
         // 지하철_노선_생성_요청
-        Map<String, String> updateRequestMap = createLineAndStationMapHelper(
+        Map<String, String> updateRequestMap = LineSteps.createLineAndStationMapHelper(
                 registeredLine.getName(),
                 registeredLine.getColor(),
                 registeredLine.getUpStationId(),
                 registeredLine.getDownStationId(),
                 registeredLine.getDistance());
         updateRequestMap.put("color", "yellogreen");
-        ExtractableResponse<Response> response  = updateLineHelper(registeredLine.getId(), updateRequestMap);
+        ExtractableResponse<Response> response  = LineSteps.updateLineHelper(registeredLine.getId(), updateRequestMap);
 
         // then
         // 지하철_노선_수정됨
@@ -228,55 +126,16 @@ public class LineAcceptanceTest extends AcceptanceTest {
         // given
         // 지하철_노선_등록되어_있음
         CreatedLineResponse registeredLine  =
-                registerLineWithStationsHelper("KangName Line", "green", "강남역","서초역", 10)
-                        .as(CreatedLineResponse.class);
+                LineSteps.registerLineWithStationsHelper("KangName Line", "green", "강남역","서초역", 10);
 
         // when
         // 지하철_노선_제거_요청
-        ExtractableResponse<Response> response  = deleteLineHelper(registeredLine.getId());
+        ExtractableResponse<Response> response  = LineSteps.deleteLineHelper(registeredLine.getId());
 
         // then
         // 지하철_노선_삭제됨
         this.assertDeleteLineSuccess(response);
     }
-
-    public static ExtractableResponse<Response> registerLineWithStationsHelper(
-            final Map<String, String> newLine,
-            final Map<String, String> upStation,
-            final Map<String, String> downStation,
-            final int distance) {
-        StationResponse upStationResponse = StationAcceptanceTest.createStationHelper(upStation).as(StationResponse.class);
-        StationResponse downStationResponse = StationAcceptanceTest.createStationHelper(downStation).as(StationResponse.class);
-        newLine.put("upStationId", upStationResponse.getId().toString());
-        newLine.put("downStationId", downStationResponse.getId().toString());
-        newLine.put("distance", String.valueOf(distance));
-        return registerLineHelper(newLine);
-    }
-
-    public static ExtractableResponse<Response> registerLineWithStationsHelper(
-            final String lineName,
-            final String lineColor,
-            final String upStationName,
-            final String downStationName,
-            final int distance) {
-        final Map newLine = createLineMapHelper(lineName,lineColor);
-        final Map upStation = StationAcceptanceTest.createStationInputHelper(upStationName);
-        final Map downStation = StationAcceptanceTest.createStationInputHelper(downStationName);
-        return registerLineWithStationsHelper(newLine, upStation, downStation, distance);
-    }
-
-    private void assertAppendNewSectionSuccess(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
-    }
-
-    private void assertDeleteSuccess(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
-    }
-
-    private void assertBadRequestFail(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
-    }
-
 
     private void assertCreateNewLineSuccess(ExtractableResponse<Response> response) {
         assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
@@ -313,135 +172,5 @@ public class LineAcceptanceTest extends AcceptanceTest {
                 map(line -> new Tuple(line.getName(), line.getColor())).
                 collect(Collectors.toList());
         assertThat(resultLines).isEqualTo(inputLines);
-    }
-
-    public static  ExtractableResponse<Response> getLinesHelper() {
-        return RestAssured.
-                given().
-                    log().all().
-                    contentType(MediaType.APPLICATION_JSON_VALUE).
-                when().
-                    get("/lines").
-                then().
-                    log().all().
-                    extract();
-    }
-
-    public static ExtractableResponse<Response> registerLineHelper(final Map<String, String> line) {
-        return RestAssured.
-                given().
-                    log().all().
-                    body(line).
-                    contentType(MediaType.APPLICATION_JSON_VALUE).
-                when().
-                    post("/lines").
-                then().
-                    log().all().
-                    extract();
-    }
-
-    public static  ExtractableResponse<Response> findLineHelper(final CreatedLineResponse inputLine) {
-        return RestAssured.
-                given().
-                    log().all().
-                    contentType(MediaType.APPLICATION_JSON_VALUE).
-                when().
-                    pathParam("lineId", inputLine.getId()).
-                    get("/lines/{lineId}").
-                then().
-                    log().all().
-                    extract();
-    }
-
-    public static ExtractableResponse<Response> updateLineHelper(final Long lineId, final Map<String, String> line) {
-        return RestAssured.
-                given().
-                    log().all().
-                    contentType(MediaType.APPLICATION_JSON_VALUE).
-                    body(line).
-                when().
-                    pathParam("lineId", lineId).
-                    patch("/lines/{lineId}").
-                then().
-                    log().all().
-                    extract();
-    }
-
-    public static ExtractableResponse<Response> deleteLineHelper(final Long lineId) {
-        return RestAssured.
-                given().
-                    log().all().
-                    contentType(MediaType.APPLICATION_JSON_VALUE).
-                when().
-                    pathParam("lineId", lineId).
-                    delete("/lines/{lineId}").
-                then().
-                    log().all().
-                    extract();
-    }
-
-    public static ExtractableResponse<Response> registerSectionHelper(final Long lineId, final Map newSection) {
-        return RestAssured.given().log().all().
-                    contentType(MediaType.APPLICATION_JSON_VALUE).
-                when().
-                    body(newSection).
-                    pathParam("lineId", lineId).
-                    post("/lines/{lineId}/sections").
-                then().
-                    log().all().
-                    extract();
-    }
-
-    public static ExtractableResponse<Response> deleteStationHelper(final Long lineId, final Long stationId) {
-        return RestAssured.given().log().all().
-                    contentType(MediaType.APPLICATION_JSON_VALUE).
-                when().
-                    pathParam("lineId", lineId).
-                    queryParam("stationId", stationId).
-                    delete("/lines/{lineId}/sections").
-                then().
-                    log().all().
-                    extract();
-    }
-
-    public static Map createLineAndStationMapHelper(
-            final String lineName,
-            final String lineColor,
-            final Long upStationId,
-            final Long downStationId,
-            final int distance){
-        Map lineMap =  new HashMap<String, String>() {
-            {
-                put("name", lineName);
-                put("color", lineColor);
-            }
-        };
-        lineMap.putAll(createSectionMapHelper(upStationId, downStationId, distance));
-        return lineMap;
-    }
-
-    public static Map createLineMapHelper(
-            final String lineName,
-            final String lineColor){
-        return new HashMap<String, String>() {
-            {
-                put("name", lineName);
-                put("color", lineColor);
-            }
-        };
-    }
-
-    public static Map createSectionMapHelper(
-            final Long upStationId,
-            final Long downStationId,
-            final int distance
-    ){
-        return new HashMap<String, String>() {
-            {
-                put("upStationId", String.valueOf(upStationId));
-                put("downStationId", String.valueOf(downStationId));
-                put("distance", String.valueOf(distance));
-            }
-        };
     }
 }
